Implement signing/verification methods directly for key types
